NDTV UI and Weather Comparator

+++ Compares Weather Assignment/src/main/java/com/pojo/weather/Main.java
@@ -23,7 +23,7 @@ public class Main {
     private Integer pressure;
     @SerializedName("humidity")
     @Expose
-    private Integer humidity;
+    private Double humidity;
 
     public Double getTemp() {
         return temp;
@@ -65,11 +65,11 @@ public class Main {
         this.pressure = pressure;
     }
 
-    public Integer getHumidity() {
+    public Double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(Integer humidity) {
+    public void setHumidity(Double humidity) {
         this.humidity = humidity;
     }
